RTP profile for audio and video conferences with minimal control. This memorandum is a revision of RFC 1890 in preparation for advancement from Proposed Standard to Draft Standard status. Readers are encouraged to use the PostScript form of this draft to see where changes from RFC 1890 are marked by change bars.
"G722 is specified in ITU-T Recommendation G.722, "7 kHz audio-coding within
64 kbit/s". The G.722 encoder produces a stream of octets, each of which
SHALL be octet-aligned in an RTP packet. The first bit transmitted in the
G.722 octet, which is the most significant bit of the higher sub-band
sample, SHALL correspond to the most significant bit of the octet in the RTP
packet.

Even though the actual sampling rate for G.722 audio is 16000 Hz, the RTP
clock rate for the G722 payload format is 8000 Hz because that value was
erroneously assigned in RFC 1890 and must remain unchanged for backward
compatibility. The octet rate or sample-pair rate is 8000 Hz."




git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@2523 d0543943-73ff-0310-b7d9-9358b9ac24b2