It appears that we should be able to include strings.h only, except msvc does not have strings.h:
and at least some linux implementations do not work unless you include both, which seems to be against spec, but should not hurt.  The most correct thing to do here would be to include strings.h then string.h wrapped in autoconf detected defines, but this should work for now.  If it is still an issue going forward, we will need to add some defines for detected presence of some header files to the public header files, which should be done with care to avoid define collisions in public header files.

http://www.opengroup.org/onlinepubs/009695399/basedefs/strings.h.html
http://www.opengroup.org/onlinepubs/009695399/basedefs/string.h.html


git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@4895 d0543943-73ff-0310-b7d9-9358b9ac24b2