merge voicemail_3_improve_storage_dir.diff  patch from MODAPP-133
allows the "storage-dir" parameter to be set on a per-user basis 
  <param name="storage-dir" value="/path/to/storage" /> 

prevents mod_voicemail to always create recursively SWITCH_GLOBAL_dirs.storage_dir. 
That may be useless when the old storage-dir (or the new per user i have introduced) is used. 
In fact, the directory is anyway always checked and maked, if necessary 


git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@9375 d0543943-73ff-0310-b7d9-9358b9ac24b2