mod_callcenter: Alot of cleanup.  Added max-wait-time and max-wait-time-with-no-agent param to a queue.
max-wait-time : allow to make the caller quit the queue after X number of waited second.  The second one allow
max-wait-time-with-no-agent : This will wait for a range of second with no agent before kicking the member out of the queue.  This allow safe shift change with all agents login out and back in

I've changed the Event value from CC-Name to either CC-Agent or CC-Queue.  So watchout for your current event socket apps.  Sorry about this change.

Lot to be tested in this commit.  Please use Jira to post problems, and new feature request can be added at the bottom of the wiki page.