mod_erlang_event: Support for reading erlang cookie from a file
mod_erlang_event will now try to read $HOME/.erlang.cookie if no cookie
is specified in the config file. You can specify an arbitary cookie file
via the 'cookie-file' parameter in the config file. The cookie file MUST
be readable by the user freeswitch is running as (either the owner or
root) and the file MUST NOT be readable/writeable/executable by any
other user (eg 0400 permissions). Thanks to James Aimonetti for the idea
and the original patch.