fix gcc 4.7.0 warning related to enum type
This amends commit 7bd9efc7f52a725b7ef9efb64b65873b4bc4330b.

We changed away from using unsigned here because it caused a warning
on Windows.  Using the more specific type, however, is causing a
warning on gcc-4.7 (it notices that the switch statement contains
values not present in the specified enum type).  So we're switching
back to unsigned here while keeping the type casts and hoping that
keeps Windows happy.