Allow setting compression level for dpkg-source
Apparently the Raspberry Pi blows up (out of memory) if you call
dpkg-source -z9 as it implicitly calls xz -9.  Thanks to Ken Rice for
running headlong into this.

Also note that while xz takes -0, dpkg-source won't take -z0, so we
check for this and make sure we call dpkg-source correctly.