Deal correctly with spaces in SWITCH_VERSION_REVISION_HUMAN
This affects the set_fs_ver bash function and the
./build/set-fs-version.sh script.

Previously we were failing to quote the value of
SWITCH_VERSION_REVISION_HUMAN when inserting it into configure.in.
When it contains spaces, which it often does, this results in a
malformed line in the generated configure script, which results in the
define in src/include/switch_version.h.in being left unset.