Improve upgrade path to freeswitch-all
That is to say, declare all appropriate Provides/Conflicts/Replaces
for freeswitch-all and freeswitch-all-dbg by enumerating the binary
packages that we're going to produce during the build.

Conflicts is more appropriate than Breaks here as both packages will
continue to provide the same files on an ongoing basis and we would
like to force the removal of the other package.