Improve configure handling of ODBC
This commit changes behavior such that if --disable-core-odbc-support
is provided we'll build without ODBC even if the libraries are there.

Previously we would always quietly build with ODBC support if it was
on the system.

Contrary to what was said in commit 72a804983, my 2012 commit
ffc8e81b7 did not affect the behavior of --disable-core-odbc-support.
We never recognized the flag as being different from not providing the
option at all.

What the commit did do was to cause us to fail loudly if
--enable-core-odbc-support was provided but the system libraries were
not there.  This behavior is preserved.

(That commit also caused us to potentially run certain checks twice,
which this commit resolves.)

You can also now provide --enable-core-odbc-support=optional which has
the same effect as the default behavior.

FS-6173 --resolve

Thanks-to: James Le Cuirot <chewi@aura-online.co.uk>