Drop use of libedit EL_REFRESH in fs_cli for now
Newer versions of libedit provide an EL_REFRESH mechanism for
redrawing the prompt.  We had previously written our own code to
handle this.  In theory handing this over to libedit should be better.

Unfortunately when using EL_REFRESH the output of sofia siptrace isn't
showing up.  We're working on that, but for now we're just force own
our redrawing code to be used instead.