Add conference member data to floor event
Between v1.2 and v1.4, member data was factored out. This makes it so
that one can not determine who the originator of a floor change event
is.

With this change, the meta data related to the conference member whom
initiated the floor change event is added to the event.

See FS-7136