FS-8585: [mod_commands] group_call: expand {} and <> to [] for each dial string
Channel variables in dial strings can be set through <>, {} or [].
<foo=bar> applies to every channel.
{foo=bar} applies to every channel, excluding channels delimited by :_:.
[foo=bar] only applies to the channel that follows [foo=bar].

Before this patch, group_call looked up users in the directory and
replaced {} with []. This logic does not work if the user has more than
one dial string, e.g. if multiple-registrations is set to true, and more
than two devices register under the same user.

This patch fixes the issue by stripping <> and {}, and inserting the
dial string before each individual dial string for each user. The
semantics of the "local_clobber" variable and the '|', ',' and ':_:'
delimiters are fully supported.