FS-10056: Fix modcheck.sh invokation
modcheck.sh gets called by build/Makefile when doing "make install" with
parameter "$(modulesdir)". This is fine when the install is done on a
live system. But when "make install" was called with a "$DESTDIR" then
this won't work, as the correct path would then be
"$(DESTDIR)$(modulesdir)". So add "$(DESTDIR)" in front.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>