FS-11149: Playing video files assigned a res_id not correctly updated on layout change
Regular conference members that are assigned a reservation ID in a layout
are automatically placed in that same reservation ID upon change of a layout.

However, playing files that are assigned a reservation ID are not properly
assigned to the same reservation ID upon layout change.

This patch makes the behavior consistent between the two types.