Fix FS-12002: $.FSRTC.prototype.stopPeer() missing self variable declaration
The stopPeer() method in jquery.FSRTC.js is missing a 'var self = this;'
declaration at the top of the function, which leads the wrong object being
assigned to self, and the self.peer.stop() function never being called.

This breaks proper hangup of calls using this function, such as a screenshare
call.