[CORE] Fix memory leak caused by persistence of AVPacket side_data
When encoded frames cannot be written to session, the loop breaks and the remaining data cannot be processed in consume_h264_bitstream function.
This lead to situation when AVPacket is not wiped out and its side_data are not released from memory causing the leak.

With this fix we are not breaking the loop when we cannot write the encoded frame to session, giving the chance to process the whole frame and in consequence to wipe out the AVPacket and free side_data.