[CORE] Fix video disruptions caused by dropping entire frame when only single packet is missed
Currently when single packet is missed and there is subsequent packet already available in buffer, which belongs to the same frame, or missed packet is marked, the entire frame is dropped.
Together with buffer resync, that causes video disruptions.

With this fix, we don't drop any frame, just because of loss of single packet. We carry on and pass the broken frame to decoder, which does all it's best to conceals errors.
Because one step back we've requested new keyframe, video will recover anyway in a moment and there will be smooth transition from the bad state we've encountered.