Reworked hashing code in previous fix for issue #694 to use ssl routines, rather than Windows specific code. Approach still works the same:
New authentication-hash entry in the event_socket_conf.xaml takes precedence over password entry and has the form $algId$salt$hash were algId is 1 for MD5, 5 for SHA256 and 6 for SHA512, salt is the value prepended to the data being hashed to add randomness and hash is the value generated when using the appropriate algorthm and salt with a specified password.

If no authentication data (hash OR password) is provided in the config file then the module will NOT listen on the event socket port.