ENGDESK-13696 Fix crypto overwriting
This fixes a problem at line 1195 in switch_core_media.c where it is checked if key is set.
It is a bug that check is done against first byte of the raw key content, because the key can start with 0 byte.
This changes the check to be done on textual representation of crypto key.