TEL-5189 Move check for telnyx_rtp_poll_timeout_s to switch_rtp_new
This avoids changing of switch_rtp_new signature, otherwise that would
require patching all the places where this method is called.
Worse, a check for channel var would have to be made at all those places,
that is redundant - moving a check into switch_rtp_new handles it all
in a single place.