Handle AMR-WB codec negotiation
Add two params to AMR-WB module, which can be set from XML config
to control codec selection based on fmtp.
Variables are:
	invite_prefer_oa
	invite_prefer_be
If set, and if multiple AMR-WB codecs are offered, then those which do not match
prefered mode (oa for octet-aligned, be for bandwidth-efficient) are excluded
from negotiation. Codec that is going to be selected is first that will match,
or last - if no match was found (last codec will match regardless of preferation).