[mod_callcenter] Fix stale members in database
When a channel is originated to an agent but the
member fails to bridge to that agent, the database is
not updated with the member status and a stale entry
of 'Answered' persists until mod_callcenter is restarted.

Additionally, cc_agent_found is set before the bridge,
therefore ending the while loop on the member channel.
If there is a problem with the agent bridge, the call is
terminated prematurely.

In this commit, we:

* Move the SQL update of the member
to the 'Answered' state to the agent thread instead of
the member's thread, so that correct data is populated.

* Reset the members state accordingly to Abandoned or
Waiting if the channels fail to bridge.

* Use cc_agent_bridged to end the member loop,
so that a member is put back on queue if the agent
channel fails to bridge.