[mod_voicemail] allow mod_voicemail to work with mod_http_cache as storage location
This commit makes mod_voicemail assume file system operations succeed if the
filename is a url. This means that you can set the `storage-dir` parameter
to a mod_http_cache url and it should just work.

Note: this does not work for deleting voicemail recordings from the http
server. Doing so would require adding delete support to mod_http_cache and
the generic file handling code.