[core] Add support for multiple 'or' values in time conditions
This commit introduces using a pipe (`|`) delimiter to match multiple 'or' values in time conditions. This enhancement will make creating common and complex time condition scenarios much simpler and easier.

For instance, to match an extension on Monday, Wednesday, and Friday during the hours 9-12 and 13-17, Up until now, you would need multiple conditions for every day and for every hour range on that day. Now, we can have something simple like this.

```xml
<extension name="ext_time_cond" continue="true">
	<condition field="destination_number" expression="^201$"/>
	<condition wday="2|4|6" hour="9-11|13-16">
		<action application="bridge" data="user/201@default"/>
	</condition>
</extension>
```