[mod_dialplan_xml] new attribute to not break from nested conditions
The default behavior is to always break from proceeding to nested conditions even if a condition matches, besides when `break="never"`.

Common sense says when a condition matches and `break` is set to `on-true` to proceed to nested conditions, and `break` should only apply to the subsequent conditions.

This commit introduces a new extension attribute, `break-from-nested` that can be set to `false` to not break from nested conditions if the condition matches; besides, if `break="on-false"`

The default behavior will be `break-from-nested=true`, so it won't break any existing systems.

Example:
```xml
<extension name="break-from-nested-test" continue="true" break-from-nested="false">
	<condition>
		<action application="set" data="var1=true" inline="true"/>
	</condition>

	<condition field="${var1}" expression="^true$" break="on-true">
		<action application="log" data="err external condition" inline="true"/>

		<condition>
			<action application="log" data="err I will only log if break-from-nested is set to false" inline="true"/>
		</condition>
	</condition>

	<condition>
		<action application="log" data="err I will only log if previous conditions did not break" inline="true"/>
	</condition>
</extension>
```