[core/media] allow transcoding between different rates in Opus
In vars.xml , when you use codec settings like this:
`<X-PRE-PROCESS cmd="set" data="internal_codec_prefs=OPUS@16000h@1c,PCMU"/>`

The switch_core_media.c line 5520 only takes the OPUS default rate (48k)
instead of 16k in the config.

Because of this, FS won't be able to choose opus if your client asks for 16k
opus. You will see same log symptom as #2226

We've found that the patch #2582 has the side effect of causing calls that
fail to negotiate the same codec as Leg A for Leg B to fail with SIP
488 INCOMPATIBLE_DESTINATION. That's not an option for us, we need backwards
compatibility with other codecs.

Rather than apply this for all codecs it should only apply for Opus. Now
calls using Opus and other codecs establish correctly on both legs, and there
are no issues with codec negotiation when the codecs on Leg A and Leg B don't
match.

Tested-by: Jérôme Poulin <jeromepoulin@gmail.com>
Co-authored-by: magiclin99 <magiclin99@gmail.com>