Add support for presence_disable_early in sofia_dialog_probe_callback
Addresses scenario where a SUBSCRIBE messages triggers a presence probe while a call with the presence_disable_early variable set or on a profile with the flag PFLAG_PRESENCE_DISABLE_EARLY set is in the early state.

This changes the behavior to skip the relevant channel when the profile has the flag set or the channel has the variable set, so it will not send the NOTIFY with the early state in response to the presence probe.